src/requirements_app.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
google-genai
python-dotenv
matplotlib
seaborn
requests
